Remove python_compat.h
We don't need compatibility with older Python, so we can
simplify the code and just move it into the C++ file where
it is used.
